stats: replace Func interface with more specialized Dist interface
The Func interface always had a tension between things you wanted to
do with statistical distributions and things you wanted to do with
mathematical functions in general. Replace it with a Dist interface
that's specific to statistical distributions. Now the representation
of a mathematical function is just a float64 -> float64 function,
which includes most of the methods of the Dist interface.