stats: always report the U statistic for the first sample
Currently, MannWhitneyUTest reports the smaller of the two U
statistics. While this is the value of U used to compute the p-value,
it means the caller can't determine which of the two samples the
returned U statistic applies to.

Instead, always return the U statistic of the first sample and
internally use the smaller of the two to compute the p-value.