stats: test U distribution against slow reference implementation
This adds several tests of the U distribution using a slow but direct
implementation as a reference implementation. This reference
implementation is based off the current implementation of the U
distribution for ties, so it's not really testing much right now (the
hand-coded reference tables make sure we're really getting it right),
but will make it possible to optimize the implementation with a high
degree of certainty.