stats: reduce calls to twoUmin and twoUmax
Currently makeUmemo calls twoUmin and twoUmax in the inner loop of
both passes. However, in the second pass these bounds are almost
always used only to determine that the value will be in the memo
table. It's much faster to just check the memo table, which is very
likely. In the few cases where it's not in the memo table, we only
have to compute twoUmax.

name       old mean              new mean              delta
UDistTies  6.82ms × (1.00,1.01)  5.89ms × (0.99,1.01)  -13.73% (p=0.000 n=10+9)