stats: simplify KDE bandwidth interface
Now the caller has two options: use the default bandwidth estimator,
or provide a fixed bandwidth. The bandwidth estimators are provided as
simple functions that take exactly the interface needed by that
estimator.

This eliminates the special support for producing KDEs from histograms
with the intent that the caller can use StreamStats to explicitly
provide a bandwidth and a (yet unwritten) function to produce a sample
approximating the histogram.