stats: make KDE type a distribution
Currently constructing a KDE is a two-step process that requires
constructing a KDE configuration, then providing a Sample. This makes
for a somewhat awkward interface because the KDE type is actually just
the configuration and the actual KDE is an opaque Dist value.

Thanks to recent simplifications to this API, we can now change this
so that the KDE simply includes a field for the Sample and a *KDE is a
Dist. With this change, there is only one type involved in a KDE and
constructing a KDE is a one step process.