scale: expand tick selection control
Currently, the only constraint a caller can place on ticks is the
maximum number of ticks. This isn't very flexible. For example, it
would be useful to be able to test render a set of ticks and require a
minimum distance between the tick labels.

Fix this by defining a TickOptions structure with far more control
over tick selection constraints, including a general predicate
function.

This is an API change.