scale: make tick levels useful and explicit
Currently tick level selection is essentially black box. As a result,
it's difficult for a caller to affect the optimization in interesting
ways. The only way to do this now is to pass a predicate function down
into the optimizer, but that's fairly limited.

Swap the role of the optimizer and the scale: make the scale API
expose ticks at any level, and make the optimizer a (useful) public
API that takes a scale and computes a tick level. The existing Ticks
method remains, but now callers that need more sophisticated
optimization can invoke the optimizer directly. This also eliminates
the need for the predicate since the caller can easily tweak the tick
level returned by the optimizer according to whatever criteria it has
and then pass the updated tick level back into the scale.