graph/graphalg: add node marking utility
The adds a facility for marking nodes that replaces math/big.Int with
a somewhat simpler, more targeted structure. Also, I actually paid
attention this time and when the graph is small, this actually stays
on the stack instead of claiming in the comments that it does, but
actually escaping to the heap.